+++ projects/frontend/CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Added tag in theme to force transition style
 - Added dynamic change of volume and audio output in menu
 - Added ubuntu_condensed.ttf as fallback font
 - Added clock in main menu